Przeniesienie getch() do osobnego pliku
