+++ getch.h
@@ -1,0 +1,8 @@
+#ifndef _GH_NIRCEK_LABIRYNT_GETCH_H_
+#define _GH_NIRCEK_LABIRYNT_GETCH_H_
+
+#ifndef _WIN32
+char getch();
+#endif
+
+#endif
