+++ main.cpp
@@ -5,39 +5,7 @@
 #include <fstream>
 #include <cmath>
 #include <vector>
-
-
-// code from https://gist.github.com/Nircek
-// under MIT license
-#include <iostream>
-#ifdef _WIN32
-#include <conio.h>
-#else
-#include <unistd.h>
-#include <termios.h>
-char getch(){
-    // src: https://stackoverflow.com/a/16361724/6732111
-    char buf=0;
-    struct termios old={0};
-    std::cout.flush();
-    if(tcgetattr(0, &old)<0)
-        throw "tcsetattr()";
-    old.c_lflag&=~ICANON;
-    old.c_lflag&=~ECHO;
-    old.c_cc[VMIN]=1;
-    old.c_cc[VTIME]=0;
-    if(tcsetattr(0, TCSANOW, &old)<0)
-        throw "tcsetattr ICANON";
-    if(read(0,&buf,1)<0)
-        throw "read()";
-    old.c_lflag|=ICANON;
-    old.c_lflag|=ECHO;
-    if(tcsetattr(0, TCSADRAIN, &old)<0)
-        throw "tcsetattr ~ICANON";
-    //std::cout<<buf;
-    return buf;
- }
-#endif
+#include "getch.h"
 
 using namespace std;
